+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.0.11 (14-01-2023)
+
+#### Improvements
+- Put showid search results on top
+- Change missing directory log level to debug
+
+#### Fixes
+- Fix parsing for shows that have year in the title
+
+-----
+
 ## 1.0.10 (15-12-2022)
 
 #### Improvements